refactor: global -> error 패키지를 exception 패키지로 이름 변경
