+++ src/main/java/org/colcum/admin/domain/post/application/PostService.java
@@ -26,10 +26,10 @@ import org.colcum.admin.domain.post.domain.type.SearchType;
 import org.colcum.admin.domain.user.dao.UserRepository;
 import org.colcum.admin.domain.user.domain.UserEntity;
 import org.colcum.admin.domain.user.domain.vo.Bookmark;
-import org.colcum.admin.global.Error.CommentNotFoundException;
-import org.colcum.admin.global.Error.EmojiNotFoundException;
-import org.colcum.admin.global.Error.InvalidAuthenticationException;
-import org.colcum.admin.global.Error.PostNotFoundException;
+import org.colcum.admin.global.exception.CommentNotFoundException;
+import org.colcum.admin.global.exception.EmojiNotFoundException;
+import org.colcum.admin.global.exception.InvalidAuthenticationException;
+import org.colcum.admin.global.exception.PostNotFoundException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
