+++ src/main/java/org/colcum/admin/global/exception/CommentNotFoundException.java
@@ -1,4 +1,4 @@
-package org.colcum.admin.global.Error;
+package org.colcum.admin.global.exception;
 
 public class CommentNotFoundException extends RuntimeException {
