+++ src/main/java/org/colcum/admin/global/exception/InvalidAuthenticationException.java
@@ -1,4 +1,4 @@
-package org.colcum.admin.global.Error;
+package org.colcum.admin.global.exception;
 
 public class InvalidAuthenticationException extends RuntimeException {
