+++ src/test/java/org/colcum/admin/domain/post/api/PostControllerTest.java
@@ -19,8 +19,8 @@ import org.colcum.admin.domain.post.domain.type.PostStatus;
 import org.colcum.admin.domain.post.domain.type.SearchType;
 import org.colcum.admin.domain.user.domain.UserEntity;
 import org.colcum.admin.domain.user.domain.vo.Bookmark;
-import org.colcum.admin.global.Error.InvalidAuthenticationException;
-import org.colcum.admin.global.Error.PostNotFoundException;
+import org.colcum.admin.global.exception.InvalidAuthenticationException;
+import org.colcum.admin.global.exception.PostNotFoundException;
 import org.colcum.admin.global.auth.WithMockJwtAuthentication;
 import org.colcum.admin.global.auth.jwt.JwtAuthentication;
 import org.colcum.admin.global.common.AbstractRestDocsTest;
