+++ src/test/java/org/colcum/admin/domain/post/application/PostServiceTest.java
@@ -26,8 +26,8 @@ import org.colcum.admin.domain.post.domain.type.SearchType;
 import org.colcum.admin.domain.user.dao.UserRepository;
 import org.colcum.admin.domain.user.domain.UserEntity;
 import org.colcum.admin.domain.user.domain.vo.Bookmark;
-import org.colcum.admin.global.Error.CommentNotFoundException;
-import org.colcum.admin.global.Error.PostNotFoundException;
+import org.colcum.admin.global.exception.CommentNotFoundException;
+import org.colcum.admin.global.exception.PostNotFoundException;
 import org.colcum.admin.global.util.Fixture;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
